main_test.go
package main

import (
	"context"
	"errors"
	"testing"

	"github.com/aws/aws-sdk-go-v2/aws"
	"github.com/aws/aws-sdk-go-v2/service/s3"
	"github.com/aws/aws-sdk-go-v2/service/s3/types"
)

// MockS3Client is a mock implementation of the S3 client
type MockS3Client struct {
	ListObjectsV2Func func(ctx context.Context, input *s3.ListObjectsV2Input, opts ...aws.Option) (*s3.ListObjectsV2Output, error)
}

func (m *MockS3Client) ListObjectsV2(ctx context.Context, input *s3.ListObjectsV2Input, opts ...aws.Option) (*s3.ListObjectsV2Output, error) {
	return m.ListObjectsV2Func(ctx, input, opts...)
}

func TestListObjects(t *testing.T) {
	t.Run("Successful Object Listing", func(t *testing.T) {
		mockClient := &MockS3Client{
			ListObjectsV2Func: func(ctx context.Context, input *s3.ListObjectsV2Input, opts ...aws.Option) (*s3.ListObjectsV2Output, error) {
				return &s3.ListObjectsV2Output{
					Contents: []types.Object{
						{Key: aws.String("object1")},
						{Key: aws.String("object2")},
					},
				}, nil
			},
		}
	})

	t.Run("Empty Bucket", func(t *testing.T) {
		// ... similar to the previous test case, but with an empty Contents slice
	})

	t.Run("Single Object", func(t *testing.T) {
		// ... similar to the previous test cases, but with a single object in the Contents slice
	})

	t.Run("Invalid Bucket Name", func(t *testing.T) {
		mockClient := &MockS3Client{
			ListObjectsV2Func: func(ctx context.Context, input *s3.ListObjectsV2Input, opts ...aws.Option) (*s3.ListObjectsV2Output, error) {
				return nil, errors.New("invalid bucket name")
			},
		}

		// ... call the function with the mock client and assert the error handling
	})

	// ... other test cases for access denied, network errors, pagination, etc.
}
